Refactor Silero VAD pretrained model loading: modularize `load` logic, unify reference model handling, and streamline imports.
