+++ crates/bunsen/src/kits/speech/silero_vad/blocks/mod.rs
@@ -7,6 +7,3 @@ mod module;
 pub use collection::*;
 #[doc(inline)]
 pub use module::*;
-
-#[cfg(any(test, feature = "testing"))]
-pub mod reference;
